entries/serializers.py
from rest_framework import serializers

from .models import Entry


class EntrySerializer(serializers.ModelSerializer):
    class Meta:
        model = Entry
        fields = ('pk', 'site_name', 'site_url', 'login', 'password')
