requirements.txt
Django==1.11.2
djangorestframework==3.6.3
factory-boy==2.11.1
